Remove node::Builder
Use Node::new() with NodeConfig instead.
